requirements.txt
streamlit
python-dotenv
plotly
pypdf2
google-generativeai
